Added Some Requirements and Handlers for Authorization Policy
Renamed MinimumAgeRequirement.cs to MinimumAgeRequirementHandler.cs since it
holds both the requirement and its handler.
Added a office badge requirement to simulate entry. It was alone because it is
supported by two handlers. One handler simulates people with permanent badges.
The other handler is to simulate temporary badges, like for visitors or staff
that did not have their badge that day. Whereas the
MinimumAgeRequirementHandler included the handler, which was then automatically
discovered, the separate handlers are added to DI as singletons.
Added a claim to simulate a permanent badge.